Move map data to it's own file
