+++ src/map/helpers/zone_generation/create_tile_layout.rs
@@ -6,6 +6,8 @@ use bevy::{
 use bevy_ecs_tilemap::map::TilemapSize;
 use rand::Rng;
 
+use super::map_data::MapData;
+
 pub fn create_hub(map_size: TilemapSize, hub_size: TilemapSize) -> MapData {
     let mut map_data = MapData::new_with_grass(map_size);
 
@@ -22,49 +24,6 @@ pub fn create_hub(map_size: TilemapSize, hub_size: TilemapSize) -> MapData {
     add_hub_entrance(&mut map_data, hub_bounds);
 
     map_data
-}
-
-pub struct MapData {
-    pub tiles: Vec<Vec<TileType>>,
-    pub colliders: Vec<EnvironmentalMapCollider>,
-}
-
-impl MapData {
-    fn new(size: TilemapSize) -> Self {
-        Self {
-            tiles: vec![vec![TileType::Ground; size.y as usize]; size.x as usize],
-            colliders: Vec::new(),
-        }
-    }
-    fn new_with_grass(size: TilemapSize) -> Self {
-        Self {
-            tiles: vec![vec![TileType::Grass; size.y as usize]; size.x as usize],
-            colliders: Vec::new(),
-        }
-    }
-    fn add_wall_collider(&mut self, start: (u32, u32), is_horizontal: bool, length: u32) {
-        let start_pos = Vec2::new(start.0 as f32, start.1 as f32);
-        let length = length as f32;
-
-        let (width, height) = if is_horizontal {
-            (length, 1.0)
-        } else {
-            (1.0, length)
-        };
-
-        let collider_pos = if is_horizontal {
-            Vec2::new(start_pos.x + (width / 2.0), start_pos.y + 0.5)
-        } else {
-            Vec2::new(start_pos.x + 0.5, start_pos.y + (height / 2.0))
-        };
-
-        self.colliders.push(EnvironmentalMapCollider {
-            collider_type: EnvironmentalType::Wall,
-            transform: Transform::from_xyz(collider_pos.x, collider_pos.y, 1.0),
-            width,
-            height,
-        });
-    }
 }
 
 pub fn create_map_with_exterior_walls_and_dead_zones(
