+++ src/map/helpers/zone_generation/mod.rs
@@ -1,7 +1,7 @@
-mod add_colliders;
 mod create_tile_layout;
 mod hub_generator;
 mod instance_generator;
+mod map_data;
 
 pub use hub_generator::generate_hub_layout;
 pub use instance_generator::generate_instance_layout;
